Refactor core errors into SRP modules
